requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cloudinary==1.29.0
coverage==6.4.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.13
django-crispy-forms==1.14.0
django-phonenumber-field==6.3.0
django-summernote==0.8.20.0
gunicorn==20.1.0
phonenumbers==8.12.50
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
